+++ packages/keyring-sdk/src/index.ts
@@ -1,3 +1,4 @@
 export * from './keyring-account-registry';
+export * from './migration';
 export * from './mnemonic';
 export * from './eth';